lock user through UI
