+++ src/api/routers/managementRouter.ts
@@ -2,9 +2,8 @@ import express, { Response } from 'express';
 import { z } from 'zod';
 
 import { isSuperUserCheck } from '../middleware/userRoleMiddleware';
+import { getAllUsersList, updateUserLock } from '../services/managementService';
 import { ParticipantRequest } from '../services/participantsService';
-import { UserService } from '../services/userService';
-import { getAllUsersList } from '../services/usersService';
 
 const handleGetAllUsers = async (req: ParticipantRequest, res: Response) => {
   const userList = await getAllUsersList();
@@ -12,10 +11,10 @@ const handleGetAllUsers = async (req: ParticipantRequest, res: Response) => {
 };
 
 const handleChangeUserLock = async (req: ParticipantRequest, res: Response) => {
-  const userService = new UserService();
   const { userId } = z.object({ userId: z.coerce.number() }).parse(req.params);
   const { isLocked } = z.object({ isLocked: z.boolean() }).parse(req.body);
-  await userService.updateUserLock(req, userId, isLocked);
+  await updateUserLock(req, userId, isLocked);
+  return res.status(200).end();
 };
 
 export function createManagementRouter() {
