+++ src/api/services/usersService.ts
@@ -198,9 +198,3 @@ export const inviteUserToParticipant = async (
     await createUserInPortal(userPartial, participant!.id, userRoleId);
   }
 };
-
-// should this be moved somewhere for only superuser actions?  managementService?
-export const getAllUsersList = async () => {
-  const userList = await User.query().where('deleted', 0).orderBy('email');
-  return userList;
-};
